+++ Modules/Popup/Database/Migrations/2017_07_03_110439_create_popup_table.php
@@ -13,23 +13,25 @@ class CreatePopupTable extends Migration
      */
     public function up()
     {
-        Schema::create('popups', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('device', 10)->default('both')->nullable();
-            $table->timestamp('begin_time')->default(null)->nullable();
-            $table->timestamp('end_time')->default(null)->nullable();
-            $table->integer('disable_hours')->default(0)->nullable();
-            $table->integer('left')->default(0)->nullable();
-            $table->integer('top')->default(0)->nullable();
-            $table->integer('height')->default(0)->nullable();
-            $table->integer('width')->default(0)->nullable();
-            $table->string('color')->nullable()->default('#000000');
-            $table->string('color_button')->nullable()->default('#393939');
-            $table->string('color_button_font')->nullable()->default('#ffffff');
-            $table->text('subject')->nullable();
-            $table->text('content')->nullable();
-            $table->tinyInteger('content_html')->default(0)->nullable();
-        });
+        if(!Schema::hasTable('popups')) {
+            Schema::create('popups', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('device', 10)->default('both')->nullable();
+                $table->timestamp('begin_time')->default(null)->nullable();
+                $table->timestamp('end_time')->default(null)->nullable();
+                $table->integer('disable_hours')->default(0)->nullable();
+                $table->integer('left')->default(0)->nullable();
+                $table->integer('top')->default(0)->nullable();
+                $table->integer('height')->default(0)->nullable();
+                $table->integer('width')->default(0)->nullable();
+                $table->string('color')->nullable()->default('#000000');
+                $table->string('color_button')->nullable()->default('#393939');
+                $table->string('color_button_font')->nullable()->default('#ffffff');
+                $table->text('subject')->nullable();
+                $table->text('content')->nullable();
+                $table->tinyInteger('content_html')->default(0)->nullable();
+            });
+        }
     }
 
     /**
@@ -39,6 +41,9 @@ class CreatePopupTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('popups');
+        if(Schema::hasTable('popups')) {
+            Schema::dropIfExists('popups');
+        }
     }
+
 }
