+++ Modules/Visit/Database/Migrations/2017_10_19_100324_create_visits_table.php
@@ -38,6 +38,8 @@ class CreateVisitsTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('visits');
+        if(Schema::hasTable('visits')) {
+            Schema::dropIfExists('visits');
+        }
     }
 }
